+++ src/program/game/entities/item/item_Potion.hpp
@@ -32,14 +32,14 @@ namespace rr {
             BIG
         } size;
 
-         Potion(Effect, Size, int am, sf::Vector2f pos = sf::Vector2f(0, 0));
+         Potion(Effect, Size, int am, sf::Vector2i pos = sf::Vector2i(0, 0));
         ~Potion();
 
         virtual void draw       (sf::RenderWindow&) override;
         virtual void update     ()                  override;
         virtual void reveal     ()                  override;
 
-        void         setPosition(sf::Vector2f)      override;
+        void         setPosition(sf::Vector2i)      override;
     };
 
 }
